+++ dev/dev.js
@@ -4,20 +4,7 @@ import plugin from '../src'
 let app = new Fpm()
 plugin.bind(app)
 let biz = new Biz('0.0.1');
-biz.addSubModules('websocket',{
-	broadcast:async function(args){
-		return new Promise( (resolve, reject) => {
-			app.publish('socketio.broadcast', args)
-            resolve({data: 1})
-		});
-	},
-	send:async function(args){
-		return new Promise( (resolve, reject) => {
-			app.publish('socketio.send', args)
-            resolve({data: 1})
-		});
-	}
-})
+biz.addSubModules('demo',{})
 app.addBizModules(biz);
 
 app.subscribe('socketio.connect', (topic, data) => {